+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -549,9 +549,9 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     final var instance2 = createInstance(nod(UUID.randomUUID()).put("hrid", "2123")).getJson();
     final var instance3 = createInstance(nod(UUID.randomUUID()).put("hrid", "12")).getJson();
     final var instance4 = createInstance(nod(UUID.randomUUID()).put("hrid", "345 12")).getJson();
-    var instance5 = createInstance(nod(id5).put("hrid", "123")).getJson();
+    createInstance(nod(id5).put("hrid", "123")).getJson();
     put(id5, marcJson);
-    instance5 = getById(id5).getJson();
+    final var instance5 = getById(id5).getJson();
 
     var response = getClient().delete(instancesStorageUrl("?query=hrid==12*"), TENANT_ID).get(10, SECONDS);
 
@@ -831,7 +831,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     assertThat(instances.size(), is(0));
     // Reports 0, not sure if this is to due with record count approximation
-    //assertThat(page.getInteger(TOTAL_RECORDS_KEY), is(5));
+    // "assertThat(page.getInteger(TOTAL_RECORDS_KEY), is(5));"
   }
 
   @Test
@@ -993,7 +993,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void cannotGetNonExistingSourceRecord() throws Exception {
+  public void cannotGetNonExistingSourceRecord() {
     getMarcJsonNotFound(UUID.randomUUID());
   }
 
@@ -1171,6 +1171,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     createHoldings(new HoldingRequestBuilder()
       .withId(expectedHoldingId)
+      .withSource(getPreparedHoldingSourceId())
       .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID)
       .forInstance(expectedInstanceId)
       .create());
@@ -1189,6 +1190,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     createHoldings(new HoldingRequestBuilder()
       .withId(otherHoldingId)
+      .withSource(getPreparedHoldingSourceId())
       .forInstance(otherInstanceId)
       .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID)
       .create());
@@ -1216,6 +1218,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     createHoldings(new HoldingRequestBuilder()
       .withId(mainLibrarySmallAngryHoldingId)
+      .withSource(getPreparedHoldingSourceId())
       .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID)
       .forInstance(smallAngryPlanetInstanceId)
       .create());
@@ -1231,6 +1234,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     createHoldings(new HoldingRequestBuilder()
       .withId(annexSmallAngryHoldingId)
+      .withSource(getPreparedHoldingSourceId())
       .withPermanentLocation(ANNEX_LIBRARY_LOCATION_ID)
       .forInstance(smallAngryPlanetInstanceId)
       .create());
@@ -1249,6 +1253,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     createHoldings(new HoldingRequestBuilder()
       .withId(nodHoldingId)
+      .withSource(getPreparedHoldingSourceId())
       .forInstance(nodInstanceId)
       .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID)
       .create());
@@ -1283,6 +1288,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     createHoldings(new HoldingRequestBuilder()
       .withId(mainLibrarySmallAngryHoldingId)
+      .withSource(getPreparedHoldingSourceId())
       .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID)
       .forInstance(smallAngryPlanetInstanceId)
       .create());
@@ -1511,17 +1517,20 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     String holdings1Uuid = loc1.toString();
     jho1.put("id", UUID.randomUUID().toString());
     jho1.put("instanceId", idJ1.toString());
+    jho1.put("sourceId", getPreparedHoldingSourceId().toString());
     jho1.put("permanentLocationId", holdings1Uuid);
 
     JsonObject jho2 = new JsonObject();
     String holdings2Uuid = loc2.toString();
     jho2.put("id", UUID.randomUUID().toString());
     jho2.put("instanceId", idJ2.toString());
+    jho2.put("sourceId", getPreparedHoldingSourceId().toString());
     jho2.put("permanentLocationId", holdings2Uuid);
 
     JsonObject jho3 = new JsonObject();
     jho3.put("id", UUID.randomUUID().toString());
     jho3.put("instanceId", idJ3.toString());
+    jho3.put("sourceId", getPreparedHoldingSourceId().toString());
     jho3.put("permanentLocationId", holdings2Uuid);
 
     createHoldings(jho1);
@@ -1887,7 +1896,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void cannotPostSynchronousBatchWithInvalidInstance() throws Exception {
+  public void cannotPostSynchronousBatchWithInvalidInstance() {
     JsonArray instancesArray = new JsonArray();
     instancesArray.add(uprooted(UUID.randomUUID()));
     instancesArray.add(smallAngryPlanet(UUID.randomUUID()).put("invalidPropertyName", "bar"));
@@ -1950,7 +1959,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void cannotPostSynchronousBatchWithDuplicateId() throws Exception {
+  public void cannotPostSynchronousBatchWithDuplicateId() {
     UUID duplicateId = UUID.randomUUID();
     JsonArray instancesArray = new JsonArray();
     instancesArray.add(uprooted(duplicateId));
@@ -2276,7 +2285,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void cannotPostSynchronousBatchWithDuplicateHrids() throws Exception {
+  public void cannotPostSynchronousBatchWithDuplicateHrids() {
     log.info("Starting cannotPostSynchronousBatchWithDuplicateHRIDs");
 
     final JsonArray instancesArray = new JsonArray();
@@ -2314,7 +2323,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void cannotPostSynchronousBatchWithHridFailure() throws Exception {
+  public void cannotPostSynchronousBatchWithHridFailure() {
     log.info("Starting cannotPostSynchronousBatchWithHRIDFailure");
 
     final JsonArray instancesArray = new JsonArray();
@@ -2547,7 +2556,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void canPostSynchronousBatchWithDiscoverySuppressedInstances() throws Exception {
+  public void canPostSynchronousBatchWithDiscoverySuppressedInstances() {
     final JsonArray instancesArray = new JsonArray();
     final UUID smallAngryPlanetId = UUID.randomUUID();
     final UUID uprootedId = UUID.randomUUID();
@@ -2569,7 +2578,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void canPostInstanceStorageBatchWithDiscoverySuppressedInstances() throws Exception {
+  public void canPostInstanceStorageBatchWithDiscoverySuppressedInstances() {
     final JsonArray instancesArray = new JsonArray();
     final UUID smallAngryPlanetId = UUID.randomUUID();
     final UUID uprootedId = UUID.randomUUID();
@@ -2734,7 +2743,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     return getCompleted.get(10, SECONDS);
   }
 
-  private void getMarcJsonNotFound(UUID id) throws Exception {
+  private void getMarcJsonNotFound(UUID id) {
     assertGetNotFound(instancesStorageUrl("/" + id + "/source-record/marc-json"));
   }
 
@@ -2999,7 +3008,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
   private Response getById(String id) {
     CompletableFuture<Response> getCompleted = new CompletableFuture<>();
-    getClient().get(instancesStorageUrl(format("/" + id)), TENANT_ID, json(getCompleted));
+    getClient().get(instancesStorageUrl(format("/%s", id)), TENANT_ID, json(getCompleted));
     try {
       return getCompleted.get(10, SECONDS);
     } catch (InterruptedException | ExecutionException | TimeoutException e) {
@@ -3007,7 +3016,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     }
   }
 
-  private void assertExists(JsonObject expectedInstance) throws Exception {
+  private void assertExists(JsonObject expectedInstance) {
     Response response = getById(expectedInstance.getString("id"));
     assertThat(response, statusCodeIs(HttpStatus.HTTP_OK));
     assertThat(response.getBody(), containsString(expectedInstance.getString("title")));