+++ src/test/java/org/folio/rest/api/ItemEffectiveLocationTest.java
@@ -72,7 +72,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
     StorageTestSuite.checkForMismatchedIds("holdings_record");
   }
 
-  public void canCalculateEffectiveLocationOnHoldingUpdate() throws Exception {
+  public void canCalculateEffectiveLocationOnHoldingUpdate() {
     UUID holdingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID);
 
     final Item[] itemsToCreate = {
@@ -88,6 +88,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
 
     JsonObject holding = holdingsClient.getById(holdingsRecordId).getJson();
     holding.put("temporaryLocationId", SECOND_FLOOR_LOCATION_ID.toString());
+    holding.put("sourceId", getPreparedHoldingSourceId().toString());
     holdingsClient.replace(holdingsRecordId, holding);
 
     for (Item item : itemsToCreate) {
@@ -97,7 +98,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void canCalculateEffectiveLocationOnHoldingRemoveTempLocationShouldBeHoldingPermLocation() throws Exception {
+  public void canCalculateEffectiveLocationOnHoldingRemoveTempLocationShouldBeHoldingPermLocation() {
     UUID holdingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID, ANNEX_LIBRARY_LOCATION_ID);
 
     final Item[] itemsToCreate = {
@@ -113,6 +114,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
 
     JsonObject holding = holdingsClient.getById(holdingsRecordId).getJson();
     holding.remove("temporaryLocationId");
+    holding.put("sourceId", getPreparedHoldingSourceId().toString());
     holdingsClient.replace(holdingsRecordId, holding);
 
     for (Item item : itemsToCreate) {
@@ -122,7 +124,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void canCalculateEffectiveLocationOnHoldingUpdateWhenSomeItemsHasLocation() throws Exception {
+  public void canCalculateEffectiveLocationOnHoldingUpdateWhenSomeItemsHasLocation() {
     UUID holdingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID);
 
     Item itemWithPermLocation = buildItem(holdingsRecordId, ONLINE_LOCATION_ID, null);
@@ -135,6 +137,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
 
     JsonObject holding = holdingsClient.getById(holdingsRecordId).getJson();
     holding.put("temporaryLocationId", SECOND_FLOOR_LOCATION_ID.toString());
+    holding.put("sourceId", getPreparedHoldingSourceId().toString());
     holdingsClient.replace(holdingsRecordId, holding);
 
     // fetch items
@@ -166,7 +169,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
   @Parameters(source = ItemEffectiveLocationTestDataProvider.class,
               method = "canCalculateEffectiveLocationOnItemUpdateParams")
   public void canCalculateEffectiveLocationOnItemUpdate(
-    PermTemp holdingLoc, PermTemp itemStartLoc, PermTemp itemEndLoc) throws Exception {
+    PermTemp holdingLoc, PermTemp itemStartLoc, PermTemp itemEndLoc) {
 
     UUID holdingsRecordId = createHolding(INSTANCE_ID, holdingLoc.perm, holdingLoc.temp);
 
@@ -204,6 +207,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
     JsonObject holdingToUpdate = createdHolding.copy();
     holdingToUpdate.remove("holdingsItems");
     holdingToUpdate.remove("bareHoldingsItems");
+    holdingToUpdate.put("sourceId", getPreparedHoldingSourceId().toString());
     setPermanentTemporaryLocation(holdingToUpdate, holdingEndLoc);
     holdingsClient.replace(holdingsRecordId, holdingToUpdate);
 
@@ -235,7 +239,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void canCalculateEffectiveLocationWhenItemAssociatedToAnotherHolding() throws Exception {
+  public void canCalculateEffectiveLocationWhenItemAssociatedToAnotherHolding() {
     UUID initialHoldingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID, ANNEX_LIBRARY_LOCATION_ID);
     UUID updatedHoldingRecordId = createInstanceAndHolding(ONLINE_LOCATION_ID, SECOND_FLOOR_LOCATION_ID);
 
@@ -254,7 +258,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
-  public void canCalculateEffectiveLocationWhenItemHasPermLocationAndAssociatedToAnotherHolding() throws Exception {
+  public void canCalculateEffectiveLocationWhenItemHasPermLocationAndAssociatedToAnotherHolding() {
     UUID initialHoldingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID, ANNEX_LIBRARY_LOCATION_ID);
     UUID updatedHoldingRecordId = createInstanceAndHolding(SECOND_FLOOR_LOCATION_ID);
 
@@ -276,7 +280,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
    * Does "UPDATE item" correctly set both item.jsonb->>'effectiveLocationId' and item.effectiveLocationId?
    */
   @Test
-  public void canSetTableFieldOnItemUpdate() throws Exception {
+  public void canSetTableFieldOnItemUpdate() {
     UUID holdingsRecordId = createInstanceAndHolding(MAIN_LIBRARY_LOCATION_ID, ANNEX_LIBRARY_LOCATION_ID);
     Item item = buildItem(holdingsRecordId, ONLINE_LOCATION_ID, null);
     createItem(item);
@@ -313,7 +317,7 @@ public class ItemEffectiveLocationTest extends TestBaseWithInventoryUtil {
     }
   }
 
-  private Item getItem(String id) throws Exception {
+  private Item getItem(String id) {
     return itemsClient.getById(UUID.fromString(id)).getJson().mapTo(Item.class);
   }