+++ src/test/java/org/folio/rest/api/SampleDataTest.java
@@ -170,4 +170,6 @@ public class SampleDataTest extends TestBase {
     assertThat(array.size(), is(expectedCount));
     assertThat(body.getInteger("totalRecords"), is(expectedCount));
   }
+
+
 }